Move images to subdirectory and mention installation
Since the chrome web store actually does enforce those image sizes,
the number of images got out of hand fast. To deal with that, I
just dumped them into a directory.

Also, I added a link to the extension on the chrome web store, as
well as a link to the google docs for how to install an unpacked
extension in case someone would rather do it themselves.