fix(load): Wait for sidebar before attaching spellbook observer
The dndbeyond app is now initializing in a way that we would finish loading the character data
before the sidebar was actually placed on the screen. This would cause the observer to fail and no
longer update the spellbook as people opened healing spells. This change adds another observer, this
time to the document body, which watches for the sidebar to appear before creating the spellbook
observer.

fix #1