chore(changelog): Increment version and add a changelog
Since I need to publish a new version with the changes that were needed after the latest dndbeyond
update, I figure now would be a good time to add a changelog. That way anyone using the extension
can easily see what changes between versions without having to keep an eye on the github issues or
PRs.