Archive the project
DND Beyond changed, closing off the api call that was used to get the character sheet. The information that we need is in redux, but I didn't see a great way to tap into that state in an appropriate way so it is probably for the best to sunset this project given it appears this is not an intended use of the app. I also see they deleted the old zendesk ticket, so I guess they don't plan to fix it. If I have time, I'll open a new ticket to get it back on their radar.